chore: Update the Werkzeug package version and the python version
The Werkzeug package is outdated and for security purposes it will need
to be upgraded. The python version that the project is utilizing also
needs to be updated on the readme.

This fixes both issues by upgrading the Werkzeug package version in
the requirements.txt and updating the python version in the readme
to ensure added security for the application

[Resolves #168303980]